Fix stack corruption with complex function arguments
For a function `f` with multiple complex arguments, saving the first argument in
`RBP - 24` and continuing to evaluate a second argument which uses more stack
without adjusting the base correctly would corrupt the stack and return garbage.

See test `functions::complex_args` for a failing test case.

Even though the change is tiny, lack of persistent data structures made this
surprisingly hard to really *understand* and get right.

Scheme 1, Rust 0 - for the first time in readability.