Add support for FFI calls into C library functions!
This patch also drops the tiny amount of asm stdlib for C to make use of
existing test functions.

This is something I've wanted for so long! Woo! 🥇

The actual implementation turned out to be a lot simpler than the version in my
head because I was preparing for a general purpose register allocator. Since all
scheme functions are still called through stack, the one function called via FFI
at any time gets an exclusive access to the 6 registers making implementation
dead simple. This would be a completely different story if either the scheme
functions needed those registers or if I had to allocate the registers for the C
stdlib.