A v1 IO with ports 👷
It took me years to get to this point and I'm really *really* fucking happy! 🎉

This 86 line patch took an entire afternoon to get right, I might actually be
pretty fucking stupid. The code is a bit rough around the edges, but it works
and can be improved over time with working tests!

Ports are a simple vector of the form `(vector 'port fname fd)` and a tiny bit
of runtime hacks tie it all together. By now its clear that the core runtime
functions exposed must be really well defined and as small as possible. As far
as possible all the other utilities must be written in scheme itself to keep
things simple. A good automatic FFI layer without having to redefine wrappers on
top of C stdlib would be pretty damn great!

🚀