Rewrite most of the C runtime in Rust 🎉 🦀
Unsurprising thing to come out of Rustfest! Woo!

The crate now exports a binary and a shared library. The library is linked to
the final executable and provides most of the primitives that was written in C.

This is a significant improvement since so much code can be reused now and there
is no need to write the same functions in multiple languages. Some of the types
can now be improved, like aliasing runtime tags into something like `Object`
instead of naked i64.

cbindgen avoids a lot of boilerplate and helps in keeping function signatures
and constants in sync. This patch is a net reduction in size even with over 40
lines of generated headers.