Coalesce public API for lang module into a single function
The analyze function serves as a good place to document the different passes and
fused the several loops into one.

This makes it much easier for the callers to get the invariants and ordering
right while making it all simpler, easier to read and shorter.

The lift/lift1, rename/rename1 hack can also be removed now which is icing on
the cake.